requirements.txt
fastapi==0.115.2
uvicorn==0.34.0
pyjwt==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
python-multipart==0.0.20
httpx==0.28.1